Check if docker gpg keyring is missing before trying to install it

+++ setup.sh
@@ -10,9 +10,11 @@ install_docker() {
   sudo apt-get -qqy update
   DEBIAN_FRONTEND=noninteractive sudo -E apt-get -qqy -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" upgrade
   sudo apt-get -yy install apt-transport-https ca-certificates curl software-properties-common pwgen gnupg
-  sudo install -m 0755 -d /etc/apt/keyrings
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-  sudo chmod a+r /etc/apt/keyrings/docker.gpg
+  if [ ! -f "/etc/apt/keyrings/docker.gpg" ]; then
+    sudo install -m 0755 -d /etc/apt/keyrings
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+    sudo chmod a+r /etc/apt/keyrings/docker.gpg
+  fi
   echo \
     "deb [arch=""$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
     ""$(. /etc/os-release && echo "$VERSION_CODENAME") stable" |
